refactor: ocr 변환 결과 레코드 계층 이동
